+++ src/Node/CharacterData.php
@@ -33,6 +33,20 @@ final class CharacterData implements Implementation
     #[\Override]
     public function toString(): string
     {
-        return '<![CDATA['.$this->value.']]>';
+        $writer = new \XMLWriter;
+        /** @psalm-suppress ImpureMethodCall */
+        $writer->openMemory();
+
+        return $this->render($writer);
+    }
+
+    #[\Override]
+    public function render(\XMLWriter $writer): string
+    {
+        /** @psalm-suppress ImpureMethodCall */
+        $writer->writeCdata($this->value);
+
+        /** @psalm-suppress ImpureMethodCall */
+        return $writer->outputMemory();
     }
 }