+++ tests/AttributeTest.php
@@ -31,28 +31,4 @@ class AttributeTest extends TestCase
         $this->assertSame('foo', $attribute->name());
         $this->assertSame('bar', $attribute->value());
     }
-
-    public function testCastWithNoValue()
-    {
-        $this->assertSame(
-            'foo',
-            Attribute::of('foo')->toString(),
-        );
-    }
-
-    public function testCastEmptyAttribute()
-    {
-        $this->assertSame(
-            'foo=""',
-            Attribute::empty('foo')->toString(),
-        );
-    }
-
-    public function testCastWithValue()
-    {
-        $this->assertSame(
-            'foo="bar"',
-            Attribute::of('foo', 'bar')->toString(),
-        );
-    }
 }